# api/proto/web/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: "3"
paths:
...
    /api/v1/talk/session-clear-unread-num:
        post:
            tags:
                - Talk
            description: 会话未读数清除接口
            operationId: Talk_SessionClearUnreadNum
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TalkSessionClearUnreadNumRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TalkSessionClearUnreadNumResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/talk/session-create:
        post:
            tags:
                - Talk
            description: 会话创建接口
            operationId: Talk_SessionCreate
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TalkSessionCreateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TalkSessionCreateResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/talk/session-delete:
        post:
            tags:
                - Talk
            description: 会话删除接口
            operationId: Talk_SessionDelete
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TalkSessionDeleteRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TalkSessionDeleteResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/talk/session-detail:
        post:
            tags:
                - Talk
            description: 会话详情接口
            operationId: Talk_SessionDetail
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TalkSessionDetailRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TalkSessionDetailResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/talk/session-disturb:
        post:
            tags:
                - Talk
            description: 会话免打扰接口
            operationId: Talk_SessionDisturb
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TalkSessionDisturbRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TalkSessionDisturbResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/talk/session-list:
        post:
            tags:
                - Talk
            description: 会话列表接口
            operationId: Talk_SessionList
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TalkSessionListRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TalkSessionListResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/talk/session-top:
        post:
            tags:
                - Talk
            description: 会话置顶接口
            operationId: Talk_SessionTop
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TalkSessionTopRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TalkSessionTopResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        TalkSessionClearUnreadNumRequest:
            type: object
            properties:
                talkMode:
                    type: integer
                    format: int32
                receiverId:
                    type: integer
                    format: int32
            description: 会话未读数清除接口请求参数
        TalkSessionClearUnreadNumResponse:
            type: object
            properties: {}
            description: 会话未读数清除接口响应参数
        TalkSessionCreateRequest:
            type: object
            properties:
                talkMode:
                    type: integer
                    description: 1:私聊 2:群聊
                    format: int32
                receiverId:
                    type: integer
                    description: 对方ID
                    format: int32
            description: 会话创建接口请求参数
        TalkSessionCreateResponse:
            required:
                - id
                - talkMode
                - receiverId
                - isTop
                - isDisturb
                - isOnline
                - isRobot
                - name
                - avatar
                - remark
                - unreadNum
                - msgText
                - updatedAt
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                talkMode:
                    type: integer
                    format: int32
                receiverId:
                    type: integer
                    format: int32
                isTop:
                    type: integer
                    format: int32
                isDisturb:
                    type: integer
                    format: int32
                isOnline:
                    type: integer
                    format: int32
                isRobot:
                    type: integer
                    format: int32
                name:
                    type: string
                avatar:
                    type: string
                remark:
                    type: string
                unreadNum:
                    type: integer
                    format: int32
                msgText:
                    type: string
                updatedAt:
                    type: string
            description: 会话创建接口响应参数
        TalkSessionDeleteRequest:
            type: object
            properties:
                talkMode:
                    type: integer
                    format: int32
                receiverId:
                    type: integer
                    format: int32
            description: 会话删除接口请求参数
        TalkSessionDeleteResponse:
            type: object
            properties: {}
            description: 会话删除接口响应参数
        TalkSessionDetailRequest:
            type: object
            properties:
                talkMode:
                    type: integer
                    format: int32
                receiverId:
                    type: integer
                    format: int32
            description: 会话详情接口请求参数
        TalkSessionDetailResponse:
            required:
                - isTop
                - isDisturb
            type: object
            properties:
                isTop:
                    type: integer
                    format: int32
                isDisturb:
                    type: integer
                    format: int32
            description: 会话详情接口响应参数
        TalkSessionDisturbRequest:
            type: object
            properties:
                talkMode:
                    type: integer
                    format: int32
                receiverId:
                    type: integer
                    format: int32
                action:
                    type: integer
                    format: int32
            description: 会话免打扰接口请求参数
        TalkSessionDisturbResponse:
            required:
                - isDisturb
            type: object
            properties:
                isDisturb:
                    type: integer
                    format: int32
            description: 会话免打扰接口响应参数
        TalkSessionItem:
            required:
                - id
                - talkMode
                - receiverId
                - isTop
                - isDisturb
                - isRobot
                - name
                - avatar
                - remark
                - unreadNum
                - msgText
                - updatedAt
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                talkMode:
                    type: integer
                    format: int32
                receiverId:
                    type: integer
                    format: int32
                isTop:
                    type: integer
                    format: int32
                isDisturb:
                    type: integer
                    format: int32
                isRobot:
                    type: integer
                    format: int32
                name:
                    type: string
                avatar:
                    type: string
                remark:
                    type: string
                unreadNum:
                    type: integer
                    format: int32
                msgText:
                    type: string
                updatedAt:
                    type: string
                atMeUserCount:
                    type: integer
                    description: 该会话中 @ 我的消息条数（群聊有效，单聊为 0），前端 atMeUserCount
                    format: int32
            description: 会话列表项
        TalkSessionListRequest:
            type: object
            properties: {}
            description: 会话列表接口请求参数
        TalkSessionListResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/TalkSessionItem'
            description: 会话列表接口响应参数
        TalkSessionTopRequest:
            type: object
            properties:
                talkMode:
                    type: integer
                    format: int32
                receiverId:
                    type: integer
                    format: int32
                action:
                    type: integer
                    format: int32
            description: 会话置顶接口请求参数
        TalkSessionTopResponse:
            required:
                - isTop
            type: object
            properties:
                isTop:
                    type: integer
                    format: int32
            description: 会话置顶接口响应参数
...